# api/openapi/flowctl.yaml
openapi: "3.0.3"
info:
  title: FlowCtl API
  description: Production-grade multi-tenant workflow platform
  version: "1.0.0"
  contact:
    name: FlowCtl Team
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /auth/callback:
    get:
      tags: [auth]
      summary: OIDC callback
      operationId: oidcCallback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token pair returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: string
          format: date-time
...